Move MinerUtility class to utility module
