+++ src/pydvl/utils/utility.py
@@ -19,12 +19,12 @@ import numpy as np
 from numpy.typing import NDArray
 from sklearn.metrics import check_scoring
 
+from pydvl.utils import Dataset
 from pydvl.utils.caching import CacheStats, memcached, serialize
 from pydvl.utils.config import MemcachedConfig
-from pydvl.utils.dataset import Dataset
 from pydvl.utils.types import Scorer, SupervisedModel
 
-__all__ = ["Utility", "DataUtilityLearning"]
+__all__ = ["Utility", "DataUtilityLearning", "MinerUtility"]
 
 logger = logging.getLogger(__name__)
 
@@ -260,3 +260,60 @@ class DataUtilityLearning:
     def data(self) -> Dataset:
         """Returns the wrapped utility's :class:`~pydvl.utils.dataset.Dataset`."""
         return self.utility.data
+
+
+class MinerUtility(Utility):
+    r"""Toy Utility that is used only for testing purposes.
+
+    Consider a group of n miners, who have discovered large bars of gold.
+
+    If two miners can carry one piece of gold, then the payoff of a coalition S is:
+
+    $$
+    v(S) = \left\{\begin{matrix}
+        |S|/2 & \text{, if} & |S| \text{ is even} \\
+        (|S| - 1)/2 & \text{, if} & |S| \text{ is odd}
+    \end{matrix}\right.
+    $$
+
+    If there are more than two miners and there is an even number of miners,
+    then the core consists of the single payoff where each miner gets 1/2.
+
+    If there is an odd number of miners, then the core is empty.
+
+    Taken from: https://en.wikipedia.org/wiki/Core_(game_theory)
+
+    :param n_miners: Number of miners that participate in the game.
+    """
+
+    def __init__(self, n_miners: int, **kwargs):
+        if n_miners <= 2:
+            raise ValueError(f"n_miners, {n_miners} should be > 2")
+        self.n_miners = n_miners
+
+        x = np.arange(n_miners)[..., np.newaxis]
+        # The y values don't matter here
+        y = np.zeros_like(x)
+
+        self.data = Dataset(x_train=x, y_train=y, x_test=x, y_test=y)
+
+    def __call__(self, indices: Iterable[int]) -> float:
+        n = len(tuple(indices))
+        if n % 2 == 0:
+            return n / 2
+        else:
+            return (n - 1) / 2
+
+    def _initialize_utility_wrapper(self):
+        pass
+
+    def exact_least_core_values(self) -> Tuple[NDArray[np.float_], float]:
+        if self.n_miners % 2 == 0:
+            values = np.array([0.5] * self.n_miners)
+            subsidy = 0.0
+        else:
+            values = np.array(
+                [(self.n_miners - 1) / (2 * self.n_miners)] * self.n_miners
+            )
+            subsidy = (self.n_miners - 1) / (2 * self.n_miners)
+        return values, subsidy
